oops oops - only enable the previous commit's sed on amd64. (although the incorrect revision is harmless also.)

+++ eclass/kde.eclass
@@ -1,6 +1,6 @@
 # Copyright 1999-2004 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/eclass/kde.eclass,v 1.109 2005/01/29 13:41:35 danarmak Exp $
+# $Header: /var/cvsroot/gentoo-x86/eclass/kde.eclass,v 1.110 2005/01/29 13:42:43 danarmak Exp $
 #
 # Author Dan Armak <danarmak@gentoo.org>
 #
@@ -56,12 +56,12 @@ kde_src_unpack() {
 	# temp fix for bug #78720, until the real bug in gcc gets fixed
 	# briefly, -fvisibility-inlines-hidden is broken on amd64
 	# this only applies to kde 3.4. the grep prevents us from removing configure unnecessarily.
-#	if useq amd64; then
+	if useq amd64; then
 		if grep -- '-fvisibility=hidden -fvisibility-inlines-hidden' admin/acinclude.m4.in >/dev/null; then
 			sed -i -e 's:-fvisibility=hidden -fvisibility-inlines-hidden:-fvisibility=hidden:' admin/acinclude.m4.in
 			rm -f configure
 		fi
-#	fi
+	fi
 }
 
 kde_src_compile() {
